Move the constants to a constants.py file
